Translate CPU info
